+++ src/aleatora/speech.py
@@ -40,7 +40,7 @@ def speech(text, lang='en', slow=False, tld='com', filename=None):
 
     decoder = MP3Decoder(mp3)
     assert(decoder.num_channels == 1)
-    data = np.concatenate([np.frombuffer(chunk, dtype=np.int16).copy() for chunk in decoder]).astype(np.float) / np.iinfo(np.int16).max
+    data = np.concatenate([np.frombuffer(chunk, dtype=np.int16).copy() for chunk in decoder]).astype(np.float32) / np.iinfo(np.int16).max
     return Stream.resample(data.tolist(), decoder.sample_rate / SAMPLE_RATE).freeze()
